Parser for Tektronix awg binary files
This VI parses the Tektronix .awg settings file, following their
little-endian binary encoding scheme and displaying the results in
arrays in LabVIEW or dumping the results into a text file.

The .AWG file format is raw binary, with repeating pattern:

32 bit unsigned int --> Length in bytes of string containing parameter
name (incl. NULL)
32 bit unsigned int --> Length in bytes of binary data containing
parameter value
x bytes --> Parameter name string
y bytes --> Parameter value in raw binary
[repeat until EOF]

Refer to Tektronix manual for data structures to expect for each
parameter. The vast majority of parameters (except for the waveforms of
course!) are simply integers or strings. Here we make the following
assumptions based on the length of the parameter value.

2 bytes --> 16 bit unsigned int
4 bytes --> 32 bit unsiigned int
8 bytes --> 64 bit unsigned (double)
All other sizes --> treat as ASCII string

This VI is meant primarily for inspection of .awg files in plain text. A
proper treatment should
hard-code the expected data types based on the parameter names rather
than guessing
based on size, and would employ an ASCII-binary encoding scheme for the
waveform
data to avoid situations where arbitrary binary data may be interpreted
as a tab or a newline character.