Tektronix AWG 5014C Read AWG Binary File (Python)
Python / Numpy version of my Tektronix AWG binary settings file parser.
Stores parameters in a Python dictionary of the form
{PARAMETER_NAME, [INDEX, VALUE_BYTES, VALUE]}
where INDEX is the order in which the file was read.

VALUE_BYTES are stored because we need to keep track of the data type
(uint16, uint32 or float64) if we ever want to write a config file back
again.

As an example at the end, this script sorts the dictionary by INDEX and
prints it to screen.